requirements.txt
SQLAlchemy~=2.0.23
fastapi~=0.103.1
starlette~=0.27.0
pydantic~=2.3.0
python-dotenv~=1.0.0
redis~=4.6.0
PyJWT~=2.8.0
passlib~=1.7.4
httpx~=0.25.0
pytest~=7.4.3
requests~=2.31.0
numpy~=1.26.2
opencv-python~=4.8.1.78
alembic~=1.12.0
uvicorn~=0.25.0
Jinja2~=3.1.2
psycopg2-binary~=2.9.9
asyncpg~=0.29.0
pydantic[email]